Move sequences to procedures/Sequence.py
