+++ laser_setup/procedures/__init__.py
@@ -1,4 +1,4 @@
-from .BaseProcedure import BaseProcedure, MetaProcedure
+from .BaseProcedure import BaseProcedure
 from .FakeProcedure import FakeProcedure
 from .IVg import IVg
 from .It import It
@@ -6,3 +6,4 @@ from .IV import IV
 from .Pt import Pt
 from .Wait import Wait
 from .LaserCalibration import LaserCalibration
+from .Sequence import MetaProcedure, MainSequence
